Fix duplicate pl.UInt64 key in print_agents' dtype_codes
The dict literal listed pl.UInt64 twice (mapped to "Q", then to "B"),
so the second entry silently won -- any UInt64 column would have been
packed as a single byte instead of 8, corrupting every field after it
in the row. No currently-written column is UInt64, so this was latent,
not active. The second entry was clearly meant to be pl.UInt8 (paired
with pl.Int8: "b" just above it).

+++ utilities/UrbanPop-scripts/upop_to_exaepi.py
@@ -1747,7 +1747,7 @@ def print_agents(
         pl.Int16: "h",
         pl.UInt16: "H",
         pl.Int8: "b",
-        pl.UInt64: "B",
+        pl.UInt8: "B",
     }
     for col in agent_fields:
         dtype = df[col].dtype
